Chore: merge two test classes to one.
